correção

+++ super_trunfo.c
@@ -32,7 +32,7 @@ int main() {
     printf("Número de pontos turísticos: ");
     scanf("%d", &numeroPontosTuristicos);
 
-    float densidadePopulacional = populacao / area;
+    float densidadePopulacional = (float)populacao / area;
     float PibPerCapita = pib / populacao;
 
 
@@ -44,7 +44,7 @@ int main() {
     printf("Área: %.2f km²\n", area);
     printf("PIB: %.2f bilhões de reais\n", pib);
     printf("Número de Pontos Turísticos: %d\n", numeroPontosTuristicos);
-    printf("Densidade Populacional: %.2f hab/km²\n", (float)densidadePopulacional);
+    printf("Densidade Populacional: %.2f hab/km²\n", densidadePopulacional);
     printf("PIB per Capita: %.2f reais\n", PibPerCapita);
 
     printf("Carta 2:\n");
@@ -55,7 +55,7 @@ int main() {
     printf("Área: %.2f km²\n", area);
     printf("PIB: %.2f bilhões de reais\n", pib);
     printf("Número de Pontos Turísticos: %d\n", numeroPontosTuristicos);
-    printf("Densidade Populacional: %.2f hab/km²\n", (float)densidadePopulacional);
+    printf("Densidade Populacional: %.2f hab/km²\n", densidadePopulacional);
     printf("PIB per Capita: %.2f reais\n", PibPerCapita);
 
     return 0;
